+++ example/lib/main.dart
@@ -8,6 +8,7 @@ import 'pages/point.dart';
 import 'pages/polygon.dart';
 import 'pages/custom.dart';
 import 'pages/bigdata.dart';
+// import 'pages/debug.dart';
 
 final routes = {
   '/': (context) => const HomePage(),
@@ -18,6 +19,7 @@ final routes = {
   '/examples/Polygon Element': (context) => PolygonPage(),
   '/examples/Custom': (context) => CustomPage(),
   '/examples/Bigdata': (context) => BigdataPage(),
+  // '/examples/Debug': (context) => DebugPage(),
 };
 
 class MyApp extends StatelessWidget {